Bender Exporter Version 5.6
Essentially, this is an obsolete feature dump.  This in preparation for
Blender 2.80 which could cause many changes.  Things removed:
-  Remove the checking for, and splitting up of meshes with more than
64k vertices.
-  Flat shading both for the mesh and the entire scene.  This can be
replaced with a Split Edges modifier.
-  Generate ES6 instead of ES3 (Tower of Babel variant).

There were a few additions:
-  Separate scene level controls for the # of decimals of (Default):
       positions / shape keys (4)
       normals (3)
       UVS (3)
      vertex colors (3)
      matrix weights (2)
-  Pickable property for meshes
-  Redesign of checking for un-applied transforms on meshes with
armatures

Due to arguments changes to some functions, you will need to restart
Blender if you had a prior version.